// src/app/interfaces/tabla2Condicion7.interface.ts
export interface Tabla2Condicion7 {
    nombregrupinves: string;
    clasificaciongrupinvesA1: string;
    clasificaciongrupinvesA: string;
    clasificaciongrupinvesB: string;
    clasificaciongrupinvesC: string;
    clasificaciongrupinvesR: string;
    otrosgrup: string;
    grupinterdis: string;
    key$?: string;
}